# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Admin API
    description: GoWind Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/menus:
        get:
            tags:
                - MenuService
            description: 查询菜单列表
            operationId: MenuService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: orderBy
                  in: query
                  description: 排序条件，其语法为JSON字符串，例如：{"val1", "-val2"}。字段名前加'-'为降序，否则为升序。
                  schema:
                    type: array
                    items:
                        type: string
                - name: query
                  in: query
                  description: AND过滤参数，其语法为json格式的字符串，如：{"key1":"val1","key2":"val2"}，具体请参见：https://github.com/tx7do/go-utils/tree/main/entgo/query/README.md
                  schema:
                    type: string
                - name: or
                  in: query
                  description: OR过滤参数，语法同AND过滤参数。
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListMenuResponse'
        post:
            tags:
                - MenuService
            description: 创建菜单
            operationId: MenuService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMenuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/menus/{id}:
        get:
            tags:
                - MenuService
            description: 查询菜单详情
            operationId: MenuService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Menu'
        put:
            tags:
                - MenuService
            description: 更新菜单
            operationId: MenuService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateMenuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - MenuService
            description: 删除菜单
            operationId: MenuService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/operation-audit-logs:
        get:
            tags:
                - OperationAuditLogService
            description: 查询操作审计日志列表
            operationId: OperationAuditLogService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: orderBy
                  in: query
                  description: 排序条件，其语法为JSON字符串，例如：{"val1", "-val2"}。字段名前加'-'为降序，否则为升序。
                  schema:
                    type: array
                    items:
                        type: string
                - name: query
                  in: query
                  description: AND过滤参数，其语法为json格式的字符串，如：{"key1":"val1","key2":"val2"}，具体请参见：https://github.com/tx7do/go-utils/tree/main/entgo/query/README.md
                  schema:
                    type: string
                - name: or
                  in: query
                  description: OR过滤参数，语法同AND过滤参数。
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListOperationAuditLogResponse'
    /admin/v1/operation-audit-logs/{id}:
        get:
            tags:
                - OperationAuditLogService
            description: 查询操作审计日志详情
            operationId: OperationAuditLogService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OperationAuditLog'
...
components:
    schemas:
...
        CreateMenuRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Menu'
            description: 创建菜单 - 请求
...
        ListMenuResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Menu'
                total:
                    type: string
            description: 查询菜单列表 - 回应
        ListOperationAuditLogResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/OperationAuditLog'
                total:
                    type: string
            description: 查询操作审计日志列表 - 回应
...
        Menu:
            type: object
            properties:
                id:
                    type: integer
                    description: 菜单ID
                    format: uint32
                status:
                    enum:
                        - OFF
                        - ON
                    type: string
                    description: 菜单状态
                    format: enum
                type:
                    enum:
                        - CATALOG
                        - MENU
                        - BUTTON
                        - EMBEDDED
                        - LINK
                    type: string
                    description: 菜单类型
                    format: enum
                path:
                    type: string
                    description: 路由路径
                redirect:
                    type: string
                    description: 重定向地址
                alias:
                    type: string
                    description: 路由别名
                name:
                    type: string
                    description: 路由命名，然后我们可以使用 name 而不是 path 来传递 to 属性给 <router-link>。
                component:
                    type: string
                    description: 指向的组件
                meta:
                    $ref: '#/components/schemas/RouteMeta'
                parentId:
                    type: integer
                    description: 父节点ID
                    format: uint32
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Menu'
                    description: 子节点树
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 菜单
        OperationAuditLog:
            type: object
            properties:
                id:
                    type: integer
                    description: 操作审计日志ID
                    format: uint32
                costTime:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                    description: 操作耗时
                success:
                    type: boolean
                    description: 操作是否成功
                requestId:
                    type: string
                    description: 请求ID
                statusCode:
                    type: integer
                    description: 状态码
                    format: int32
                reason:
                    type: string
                    description: 操作失败原因
                location:
                    type: string
                    description: 操作地理位置
                operation:
                    type: string
                    description: 操作方法
                method:
                    type: string
                    description: 请求方法
                path:
                    type: string
                    description: 请求路径
                apiModule:
                    type: string
                    description: API所属模块
                apiDescription:
                    type: string
                    description: API操作描述
                referer:
                    type: string
                    description: 请求源
                requestUri:
                    type: string
                    description: 请求URI
                requestHeader:
                    type: string
                    description: 请求头
                requestBody:
                    type: string
                    description: 请求体
                response:
                    type: string
                    description: 响应信息
                userId:
                    type: integer
                    description: 操作者用户ID
                    format: uint32
                username:
                    type: string
                    description: 操作者账号名
                clientIp:
                    type: string
                    description: 操作者IP
                userAgent:
                    type: string
                    description: 浏览器的用户代理信息
                browserName:
                    type: string
                    description: 浏览器名称
                browserVersion:
                    type: string
                    description: 浏览器版本
                clientId:
                    type: string
                    description: 客户端ID
                clientName:
                    type: string
                    description: 客户端名称
                osName:
                    type: string
                    description: 操作系统名称
                osVersion:
                    type: string
                    description: 操作系统版本
                tenantId:
                    type: integer
                    description: 租户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
            description: 操作审计日志
...
        RouteMeta:
            type: object
            properties:
                activeIcon:
                    type: string
                    description: 激活图标，用于：菜单、tab
                activePath:
                    type: string
                    description: 当前激活的菜单，有时候不想激活现有菜单，需要激活父级菜单时使用
                affixTab:
                    type: boolean
                    description: 是否固定标签页
                affixTabOrder:
                    type: integer
                    description: 固定标签页的顺序
                    format: int32
                authority:
                    type: array
                    items:
                        type: string
                    description: 权限列表，需要特定的角色标识才可以访问
                badge:
                    type: string
                    description: 徽标
                badgeType:
                    enum:
                        - dot
                        - normal
                    type: string
                    description: 徽标类型
                badgeVariants:
                    enum:
                        - default
                        - destructive
                        - primary
                        - success
                        - warning
                    type: string
                    description: 徽标颜色
                hideChildrenInMenu:
                    type: boolean
                    description: 当前路由的子级在菜单中不展现
                hideInBreadcrumb:
                    type: boolean
                    description: 当前路由在面包屑中不展现
                hideInMenu:
                    type: boolean
                    description: 当前路由在菜单中不展现
                hideInTab:
                    type: boolean
                    description: 当前路由在标签页不展现
                icon:
                    type: string
                    description: 图标，用于：菜单、标签页
                iframeSrc:
                    type: string
                    description: iframe 地址
                ignoreAccess:
                    type: boolean
                    description: 忽略权限，直接可以访问
                keepAlive:
                    type: boolean
                    description: 开启KeepAlive缓存
                link:
                    type: string
                    description: 外链-跳转路径
                loaded:
                    type: boolean
                    description: 路由是否已经加载过
                maxNumOfOpenTab:
                    type: integer
                    description: 标签页最大打开数量
                    format: int32
                menuVisibleWithForbidden:
                    type: boolean
                    description: 菜单可以看到，但是访问会被重定向到403
                openInNewWindow:
                    type: boolean
                    description: 在新窗口打开
                order:
                    type: integer
                    description: 排序编号，用于路由->菜单排序
                    format: int32
                title:
                    type: string
                    description: 标题名称，路由上显示的标题
            description: 路由元数据
...
        UpdateMenuRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/Menu'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新菜单 - 请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...